// back/src/interfaces/hotelGuests.ts
export interface hotelGuest{
    Id_hospede: number;
    Hospedagem: number;
    Nome: string;
    Email: string;
    CPF: string;
    Nascimento: Date;
    Sexo: string;
    Profissao: string;
    Nacionalidade: string;
    Descricao_documento: string;
    Tipo_documento: string;
    Orgao_exp: string;
    DDI_celular: string;
    DDD_celular: string;
    Numero_celular: string;
    DDI_telefone: string;
    DDD_telefone: string;
    Numero_telefone: string;
    Endereco: string;
    Numero: string;
    Bairro: string;
    Cidade: string;
    Estado: string;
    Pais: string;
    CEP: string;
    Ultima_cidade: string;
    Ultima_estado: string;
    Ultima_pais: string;
    Proxima_cidade: string;
    Proxima_estado: string;
    Proxima_pais: string;
    Motivo: number;
    Meio_transporte: number;
    Observacoes: string;

}